Add new animation that moves from the side
Also restructered the shifting functionality a bit
